Moved elab to text_elab, started notebook_elab.
